GKE Autopilot: Add support for Extended Duration pods under a feature gate
Introduces Extended Duration Pod support for Autopilot under feature
gate "GKEAutopilotExtendedDurationPods". This feature has some rough
edges with Agones on Autopilot 1.27, so I'm introducing this as a dev
feature gate until we sort everything out.

ED pods have the same UI as disabling Autopilot on GKE Standard [1]:
annotate with `cluster-autoscaler.kubernetes.io/safe-to-evict=false`
and your pod is protected from eviction. However, Autopilot handles
node upgrades specially and just autoscales new nodes at new versions,
so the concept of "safe to evict on upgrade" is unnecessary. Since we
can support the concept and do nothing really special, do so: We pull
out the "generic" implementation of SetEviction into
cloudproduct/eviction and call that.

[1] https://cloud.google.com/kubernetes-engine/docs/how-to/extended-duration-pods

Towards #3366
