+++ pkg/cloudproduct/gke/gke_test.go
@@ -77,6 +77,7 @@ func TestSyncPodPortsToGameServer(t *testing.T) {
 
 func TestValidateGameServer(t *testing.T) {
 	for name, tc := range map[string]struct {
+		edPods      bool
 		ports       []agonesv1.GameServerPort
 		scheduling  apis.SchedulingStrategy
 		safeToEvict agonesv1.EvictionSafe
@@ -107,7 +108,7 @@ func TestValidateGameServer(t *testing.T) {
 			safeToEvict: agonesv1.EvictionSafeAlways,
 			scheduling:  apis.Packed,
 		},
-		"bad policy => fails validation": {
+		"bad policy (no feature gates) => fails validation": {
 			ports: []agonesv1.GameServerPort{
 				{
 					Name:          "best-tcpudp",
@@ -137,9 +138,39 @@ func TestValidateGameServer(t *testing.T) {
 				field.Invalid(field.NewPath("spec", "eviction", "safe"), "OnUpgrade", "eviction.safe OnUpgrade not supported on GKE Autopilot"),
 			},
 		},
+		"bad policy (GKEAutopilotExtendedDurationPods enabled) => fails validation but OnUpgrade works": {
+			edPods: true,
+			ports: []agonesv1.GameServerPort{
+				{
+					Name:          "best-tcpudp",
+					PortPolicy:    agonesv1.Dynamic,
+					ContainerPort: 4321,
+					Protocol:      agonesv1.ProtocolTCPUDP,
+				},
+				{
+					Name:          "bad-udp",
+					PortPolicy:    agonesv1.Static,
+					ContainerPort: 1234,
+					Protocol:      corev1.ProtocolUDP,
+				},
+				{
+					Name:          "another-bad-udp",
+					PortPolicy:    agonesv1.Static,
+					ContainerPort: 1234,
+					Protocol:      corev1.ProtocolUDP,
+				},
+			},
+			safeToEvict: agonesv1.EvictionSafeOnUpgrade,
+			scheduling:  apis.Distributed,
+			want: field.ErrorList{
+				field.Invalid(field.NewPath("spec", "scheduling"), "Distributed", "scheduling strategy must be Packed on GKE Autopilot"),
+				field.Invalid(field.NewPath("spec", "ports").Index(1).Child("portPolicy"), "Static", "portPolicy must be Dynamic on GKE Autopilot"),
+				field.Invalid(field.NewPath("spec", "ports").Index(2).Child("portPolicy"), "Static", "portPolicy must be Dynamic on GKE Autopilot"),
+			},
+		},
 	} {
 		t.Run(name, func(t *testing.T) {
-			causes := (&gkeAutopilot{}).ValidateGameServerSpec(&agonesv1.GameServerSpec{
+			causes := (&gkeAutopilot{useExtendedDurationPods: tc.edPods}).ValidateGameServerSpec(&agonesv1.GameServerSpec{
 				Ports:      tc.ports,
 				Scheduling: tc.scheduling,
 				Eviction:   &agonesv1.Eviction{Safe: tc.safeToEvict},
@@ -191,7 +222,7 @@ func TestPodSeccompUnconfined(t *testing.T) {
 	}
 }
 
-func TestSetSafeToEvict(t *testing.T) {
+func TestSetEvictionNoExtended(t *testing.T) {
 	emptyPodAnd := func(f func(*corev1.Pod)) *corev1.Pod {
 		pod := &corev1.Pod{
 			ObjectMeta: metav1.ObjectMeta{
@@ -260,7 +291,7 @@ func TestSetSafeToEvict(t *testing.T) {
 		},
 	} {
 		t.Run(desc, func(t *testing.T) {
-			err := (&gkeAutopilot{}).SetEviction(tc.eviction, tc.pod)
+			err := setEvictionNoExtended(tc.eviction, tc.pod)
 			if tc.wantErr {
 				assert.Error(t, err)
 				return
